// apps/web/src/lib/simulation/types.ts
export const BIOME = {
  DEEP_SEA: 0,
  SHALLOW_SEA: 1,
  SAND: 2,
  GRASS: 3,
  FOREST: 4,
  MOUNTAIN: 5,
  SNOW: 6,
} as const;
export type BiomeId = (typeof BIOME)[keyof typeof BIOME];
// ...
export const WORLD_TILES = 1024;